+++ src/main/java/edu/ie3/datamodel/models/input/thermal/ThermalUnitInput.java
@@ -7,20 +7,19 @@ package edu.ie3.datamodel.models.input.thermal;
 
 import edu.ie3.datamodel.io.extractor.HasThermalBus;
 import edu.ie3.datamodel.models.OperationTime;
-import edu.ie3.datamodel.models.input.AssetInput;
 import edu.ie3.datamodel.models.input.OperatorInput;
 import java.util.Objects;
 import java.util.UUID;
 
 /** Abstract class for grouping all common properties to thermal models. */
-public abstract class ThermalUnitInput extends AssetInput implements HasThermalBus {
+public abstract class ThermalUnitInput extends ThermalInput implements HasThermalBus {
   /** The thermal bus, a thermal unit is connected to. */
   private final ThermalBusInput thermalBus;
 
   /**
    * @param uuid Unique identifier of a certain thermal input
    * @param id Identifier of the thermal unit
-   * @param thermalBus hermal bus, a thermal unit is connected to
+   * @param thermalBus thermal bus, a thermal unit is connected to
    */
   ThermalUnitInput(UUID uuid, String id, ThermalBusInput thermalBus) {
     super(uuid, id);